requirements.txt
python-dotenv>=0.21.0
sqlalchemy>=2.0.25
pydantic>=2.5.1
fastapi>=0.103.0
python-jose>=3.3.0
passlib>=1.7.4